bulk upload now persists and sends to S3
