+++ app/models/forms/bulk_upload_lettings/upload_your_file.rb
@@ -7,6 +7,7 @@ module Forms
 
       attribute :year, :integer
       attribute :file
+      attribute :current_user
 
       validates :file, presence: true
       validate :validate_file_is_csv
@@ -27,7 +28,29 @@ module Forms
         bulk_upload_lettings_log_path(id: "checking-file", form: { year: })
       end
 
+      def save!
+        bulk_upload = BulkUpload.create!(
+          user: current_user,
+          log_type: BulkUpload.log_types[:lettings],
+          year:,
+        )
+
+        if upload_enabled?
+          storage_service.write_file(bulk_upload.identifier, File.read(file.path))
+        end
+
+        true
+      end
+
     private
+
+      def upload_enabled?
+        !Rails.env.development?
+      end
+
+      def storage_service
+        @storage_service ||= Storage::S3Service.new(Configuration::PaasConfigurationService.new, ENV["CSV_BUCKET"])
+      end
 
       def validate_file_is_csv
         return unless file
