+++ app/models/forms/bulk_upload_lettings/year.rb
@@ -27,6 +27,10 @@ module Forms
         bulk_upload_lettings_log_path(id: "prepare-your-file", form: { year: })
       end
 
+      def save!
+        true
+      end
+
     private
 
       def possible_years
